dedupe logic
